tau-core/pyproject.toml
[tool.poetry]
name = "tau-core"
version = "0.1.0"
description = "Shared core library for tau task management tools"
authors = []
readme = "README.md"
packages = [{include = "tau_core"}]

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
pycryptodome = "^3.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
